UI updates and new page updates

+++ frontend/src/helpers/axios.js
@@ -15,7 +15,7 @@ const axiosIntance =axios.create({
 axiosIntance.interceptors.request.use((req) => {
     const {auth} = store.getState();
     if(auth.token){
-        req.headers.Authorization = `Bearer ${auth.token}`;
+        req.headers.Authorization = `Bearer ${auth.token}`
     }
     return req;
 })
@@ -24,8 +24,8 @@ axiosIntance.interceptors.response.use((res) => {
     return res;
 }, (error) => {
     console.log(error.response);
-    const {status} = error.response;
-    if(status === 500){
+    const status = error.response ? error.response.status : 500;
+    if(status && status === 500){
         localStorage.clear();
         store.dispatch({type: authConstants.LOGOUT_SUCCESS})
     }
